check for wget

+++ prometheus-node-exporter.sh
@@ -4,9 +4,15 @@
 useradd node_exporter -s /sbin/nologin
 
 # Get node exporter binary and copy it into place
-wget https://github.com/prometheus/node_exporter/releases/download/v1.3.1/node_exporter-1.3.1.linux-amd64.tar.gz
-tar xvfz node_exporter-*.*-amd64.tar.gz
-cp node_exporter-*.*-amd64/node_exporter /usr/sbin/
+if [[ -f "wget" ]]; then
+    echo "Checking for wget... found"
+    wget https://github.com/prometheus/node_exporter/releases/download/v1.3.1/node_exporter-1.3.1.linux-amd64.tar.gz
+    tar xvfz node_exporter-*.*-amd64.tar.gz
+    cp node_exporter-*.*-amd64/node_exporter /usr/sbin/
+else
+    echo "wget not found, exiting"
+    exit 1
+fi
 
 # Create launchdaemon to start node exporter binary
 /bin/cat << EOF > "/etc/systemd/system/node_exporter.service"
